requirements.txt
PyYAML==6.0.1
rich>=10.0.0
